requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
motor==3.3.1
pymongo==4.5.0
pydantic==2.4.2
pydantic[email]==2.4.2
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.0
python-multipart==0.0.6
email-validator==2.0.0
